refactor(ops): improve json-joy alignment, performance, and type safety
- Simplify type/test_type matching with getTypeName + isWholeNumber
- Add HasStr field to StrDelOperation for str/len mode distinction
- Fix compact codec str_del encode/decode to match json-joy format
- Eliminate ApplyOp/ApplyOps JSON serialization roundtrip
- Change str_ins/str_del Pos/Len from float64 to int
- Remove redundant deepclone in add/replace (clone-once-at-entry)
- Optimize DefaultOptions with zero-allocation for common case
- Reduce excessive error wrapping in patch application

+++ codec/binary/encoder.go
@@ -212,7 +212,7 @@ func encodeStrIns(w *msgp.Writer, o *op.StrInsOperation) error {
 	if err := encodePath(w, o.Path()); err != nil {
 		return err
 	}
-	if err := w.WriteFloat64(o.Pos); err != nil {
+	if err := w.WriteFloat64(float64(o.Pos)); err != nil {
 		return err
 	}
 	return w.WriteString(o.Str)
@@ -226,10 +226,10 @@ func encodeStrDel(w *msgp.Writer, o *op.StrDelOperation) error {
 	if err := encodePath(w, o.Path()); err != nil {
 		return err
 	}
-	if err := w.WriteFloat64(o.Pos); err != nil {
+	if err := w.WriteFloat64(float64(o.Pos)); err != nil {
 		return err
 	}
-	return w.WriteFloat64(o.Len)
+	return w.WriteFloat64(float64(o.Len))
 }
 
 // encodeExtend encodes an extend operation: [code, path, properties, deleteNull].
